refactor(front-end): move MarketingBanner out of GetStarted into Marketing
A marketing banner doesn't belong under GetStarted/. Move it to
components/Marketing/ (alongside PremiumTooltip etc.) and update the import.
